+++ internal/index/index.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/blevesearch/bleve/v2"
 	"github.com/blevesearch/bleve/v2/mapping"
-	"github.com/pavelpivovarov/glow/internal/article"
-	"github.com/pavelpivovarov/glow/internal/config"
+	"git.netra.pivpav.com/public/glow/internal/article"
+	"git.netra.pivpav.com/public/glow/internal/config"
 )
 
 // Index handles Bleve search indexing